alligator.dat(mcexact)	R Documentation

Data frame of alligator food choice by lake gender and size.

Description

A 5x4x2x2 table of alligators food choice, lake, gender and size stored in a data frame.

Usage

data(alligator.dat)

Format

allgiator.dat is a data frame

References

Agresti (1990). Categorical Data Analysis. Wiley-Interscience. Page 308.

See Also

[Package Contents]