hyper(mcexact)	R Documentation

Generalized hypergeometric distribution on the log scale

Description

Returns the log hypergeometric conditional probability of a vector of Poisson counts up to an additive constant of proportionallity.

Usage


hyper(y)

Arguments

y 	 Counts

Details

Value

A real-valued scalar.

Note

Author(s)

Brian S. Caffo

References

See Also

mcexact

Examples



[Package Contents]