mcexact.internals(mcexact)	R Documentation

Internal functions for mcexact

Description

Internal functions used for mcexact.

Usage


rounded.tprob(y, m, s, df)
errorcheck(y,
           x,
           stat,
           dens,
           nosim,
           method,
           savechain,
           tdf,
           maxiter,
           p,
           batchsize)

Arguments

y 	 counts

m 	 means

s 	 variances

df 	 degrees of freedom

x 	 model matrix

stat 	 function to be checked

dens 	 function to be checked

nosim 	 number of simulations to be checked

method 	 method to be checked

savechain 	 savechain flag to be checked

tdf 	 t degrees of freedom to be checked

maxiter 	 maximum number of iterations to be checked

p 	 prop of table entries left fixed to be checked

batchsize 	 batchsize to be checked

Details

Value

rounded.tprob returns rounded student's t probabilities for integers y with means m and variances s. errorcheck is a function containing most of the error checking that mcexact performs.

Author(s)

Brian S. Caffo

References

See Also

mcexact

Examples



[Package Contents]