pathologist.dat(mcexact)	R Documentation

Cross-classification of 118 tumor grades by two pathologists

Description

Data frame of a 5x5 table of pathologists ratings of 118 tumors. Each pathologist, A and B, rated the tumors on a scale of 1 to 5. y represents the counts of each combination of ratings.

Usage


data(pathologist.dat)

Format

A list containing y-the counts and x-the design matrix for the uniform association model.

References

Agresti (1990). Categorical Data Analysis. Wiley-Interscience. Page 368.

See Also

Page 263 of Categorical Data Analysis for a description of the uniform assocation model.

[Package Contents]