print.mcexact(mcexact)	R Documentation

Print utilities for mcexact

Description

Generic print methods for objects of class cab and bab outputted from mcexact.

Usage


print(x,...)
summary(object,...)

Arguments

x 	 An object of class bab or cab

object 	 An object of class bab or cab

... 	 Unused, retained for generic method consistency

Details

Value

print.bab and print.cab prints the P-value estimates obtained from mcexact. summary.bab and summary.cab prints extra information and returns a matrix of the P-value estimates.

Author(s)

Brian S. Caffo

References

See Also

mcexact

Examples


#data(residence)
#resid.mcx <- mcexact(residence$y ~ residence$x, nosim = 10 ^ 2, maxiter = 10 ^ 4)
#resid.mcx #calls print.bab
#print(resid.mcx) #calls print.bab
#summary(resid.mcx) #calls summary.bab

[Package Contents]