update.bab(mcexact)	R Documentation

Update method for objects of class bab

Description

An update method for objects created by mcexact when method = 'bab'.

Usage


update(object, ...)
bab(args, nosim = NULL, maxiter = NULL, savechain = FALSE)

Arguments

object 	 Output from mcexact

... 	 Alternative arguments for the update

args 	 Output from mcexact

nosim 	 The desired number of simulations to be performed in the update

maxiter 	 The maximum number of iterations allowed.

savechain 	 Saves the chain of goodness-of-fit statistics and their importance weights

Details

The method update.bab calls the function bab, which is the engine for mcexact when method = 'bab'.

Value

A list of the form outputted from mcexact

Note

Author(s)

Brian S. Caffo

References

See Also

mcexact

Examples


data(residence.dat)
mcx <- mcexact(y ~ res.1985 + res.1980 + factor(sym.pair), data = residence.dat) 
summary(mcx)
mcx <- update(mcx, nosim = 10 ^ 4, maxiter = 10 ^ 6)
summary(mcx)

[Package Contents]