update.cab(mcexact)	R Documentation

Update method for objects of class cab

Description

An update method for objects created by mcexact when method = 'cab'.

Usage


update(object, ...)
cab(args,
    nosim = NULL,
    batchsize = NULL,
    savechain = FALSE,
    p = NULL,
    flush = FALSE)

Arguments

object 	 Output from mcexact

... 	 Alternative arguments for the update

args 	 Output from mcexact

nosim 	 The number of simulations to be performed in the update

batchsize 	 A new batchsize

savechain 	 Saves the chain of goodness-of-fit statistics

p 	 An updated proportion of simulated tables left fixed.

flush 	 Should the previous information be discarded? flush should be set to true if the batchsize is changed.

Details

The method update.cab calls the function cab, which is the engine for mcexact when method = 'cab'.

Value

A list of the form outputted from mcexact

Note

Author(s)

Brian S. Caffo

References

See Also

mcexact

Examples


data(residence.dat)
mcx <- mcexact(y ~ res.1985 + res.1980 + factor(sym.pair),
               data = residence.dat,
               method = "cab",
               p = .5,
               batchsize = 100) 
summary(mcx)
mcx <- update(mcx, nosim = 10 ^ 4)
summary(mcx)

[Package Contents]